change session detail in a component, add MentorshipSessionDetailPage
